# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PSYX MAIL GEN API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/change-password:
    post:
      operationId: changePassword
      tags: [user]
      summary: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /owner/system-logs:
    get:
      operationId: ownerGetSystemLogs
      tags: [owner]
      summary: Get system/error logs
      parameters:
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [info, warn, error]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: System logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemLogListResponse"
...
  /owner/admins:
    get:
      operationId: ownerListAdmins
      tags: [owner]
      summary: List all admin users
      responses:
        "200":
          description: Admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: ownerCreateAdmin
      tags: [owner]
      summary: Grant admin role to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRoleInput"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /owner/admins/{id}:
    delete:
      operationId: ownerRemoveAdmin
      tags: [owner]
      summary: Remove admin role from user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admin removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /owner/all-emails:
    get:
      operationId: ownerGetAllEmails
      tags: [owner]
      summary: Get all emails in the system (owner only)
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: integer
        - name: folder
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: All emails
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerEmailListResponse"
  /owner/grant-admin-by-email:
    post:
      operationId: ownerGrantAdminByEmail
      tags: [owner]
      summary: Grant admin role to any user by email (any domain)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantAdminByEmailInput"
      responses:
        "200":
          description: Admin granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /owner/maintenance:
    get:
      operationId: ownerGetMaintenanceMode
      tags: [owner]
      summary: Get maintenance mode status
      responses:
        "200":
          description: Maintenance status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceStatus"
    post:
      operationId: ownerSetMaintenanceMode
      tags: [owner]
      summary: Toggle maintenance mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceModeInput"
      responses:
        "200":
          description: Maintenance mode set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceStatus"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    UserProfile:
      type: object
      required: [id, username, email, displayName, role, status, createdAt, storageUsed, storageLimit]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
        status:
          type: string
        avatarUrl:
          type: ["string", "null"]
        recoveryEmail:
          type: ["string", "null"]
        theme:
          type: string
        language:
          type: string
        notificationsEnabled:
          type: boolean
        twoFactorEnabled:
          type: boolean
        storageUsed:
          type: integer
        storageLimit:
          type: integer
        createdAt:
          type: string
        lastLoginAt:
          type: ["string", "null"]
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        recoveryEmail:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        theme:
          type: string
        language:
          type: string
        notificationsEnabled:
          type: boolean
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
...
    OwnerEmailItem:
      type: object
      required: [id, subject, fromEmail, toEmail, folder, createdAt, userId, username]
      properties:
        id:
          type: integer
        subject:
          type: string
        fromEmail:
          type: string
        toEmail:
          type: string
        body:
          type: string
        folder:
          type: string
        isRead:
          type: boolean
        isStarred:
          type: boolean
        createdAt:
          type: string
        userId:
          type: integer
        username:
          type: string
    OwnerEmailListResponse:
      type: object
      required: [emails, total, page, limit]
      properties:
        emails:
          type: array
          items:
            $ref: "#/components/schemas/OwnerEmailItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    GrantAdminByEmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
...
    SystemLogListResponse:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/SystemLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SystemLog:
      type: object
      required: [id, level, message, createdAt]
      properties:
        id:
          type: integer
        level:
          type: string
          enum: [info, warn, error]
        message:
          type: string
        details:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AdminUser:
      type: object
      required: [id, username, email, displayName, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        createdAt:
          type: string
    AdminRoleInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
    MaintenanceStatus:
      type: object
      required: [enabled, message]
      properties:
        enabled:
          type: boolean
        message:
          type: string
    MaintenanceModeInput:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        message:
          type: string